Struct google_cloudfunctions1::api::CloudFunction

pub struct CloudFunction {
    pub available_memory_mb: Option<i32>,
    pub build_environment_variables: Option<HashMap<String, String>>,
    pub build_id: Option<String>,
    pub build_worker_pool: Option<String>,
    pub description: Option<String>,
    pub entry_point: Option<String>,
    pub environment_variables: Option<HashMap<String, String>>,
    pub event_trigger: Option<EventTrigger>,
    pub https_trigger: Option<HttpsTrigger>,
    pub ingress_settings: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub max_instances: Option<i32>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub runtime: Option<String>,
    pub service_account_email: Option<String>,
    pub source_archive_url: Option<String>,
    pub source_repository: Option<SourceRepository>,
    pub source_token: Option<String>,
    pub source_upload_url: Option<String>,
    pub status: Option<String>,
    pub timeout: Option<String>,
    pub update_time: Option<String>,
    pub version_id: Option<String>,
    pub vpc_connector: Option<String>,
    pub vpc_connector_egress_settings: Option<String>,
}

Describes a Cloud Function that contains user computation executed in response to an event. It encapsulate function and triggers configurations.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations functions create projects (request)

locations functions get projects (response)

locations functions patch projects (request)

Fields
available_memory_mb: Option<i32>

The amount of memory in MB available for a function. Defaults to 256MB.

build_environment_variables: Option<HashMap<String, String>>

Build environment variables that shall be available during build time.

build_id: Option<String>

Output only. The Cloud Build ID of the latest successful deployment of the function.

build_worker_pool: Option<String>

Name of the Cloud Build Custom Worker Pool that should be used to build the function. The format of this field is projects/{project}/locations/{region}/workerPools/{workerPool} where {project} and {region} are the project id and region respectively where the worker pool is defined and {workerPool} is the short name of the worker pool. If the project id is not the same as the function, then the Cloud Functions Service Agent (service-@gcf-admin-robot.iam.gserviceaccount.com) must be granted the role Cloud Build Custom Workers Builder (roles/cloudbuild.customworkers.builder) in the project.

description: Option<String>

User-provided description of a function.

entry_point: Option<String>

The name of the function (as defined in source code) that will be executed. Defaults to the resource name suffix, if not specified. For backward compatibility, if function with given name is not found, then the system will try to use function named "function". For Node.js this is name of a function exported by the module specified in source_location.

environment_variables: Option<HashMap<String, String>>

Environment variables that shall be available during function execution.

event_trigger: Option<EventTrigger>

A source that fires events in response to a condition in another service.

https_trigger: Option<HttpsTrigger>

An HTTPS endpoint type of source that can be triggered via URL.

ingress_settings: Option<String>

The ingress settings for the function, controlling what traffic can reach it.

labels: Option<HashMap<String, String>>

Labels associated with this Cloud Function.

max_instances: Option<i32>

The limit on the maximum number of function instances that may coexist at a given time. In some cases, such as rapid traffic surges, Cloud Functions may, for a short period of time, create more instances than the specified max instances limit. If your function cannot tolerate this temporary behavior, you may want to factor in a safety margin and set a lower max instances value than your function can tolerate. See the Max Instances Guide for more details.

name: Option<String>

A user-defined name of the function. Function names must be unique globally and match pattern projects/*/locations/*/functions/*

network: Option<String>

The VPC Network that this cloud function can connect to. It can be either the fully-qualified URI, or the short name of the network resource. If the short network name is used, the network must belong to the same project. Otherwise, it must belong to a project within the same organization. The format of this field is either projects/{project}/global/networks/{network} or {network}, where {project} is a project id where the network is defined, and {network} is the short name of the network. This field is mutually exclusive with vpc_connector and will be replaced by it. See the VPC documentation for more information on connecting Cloud projects.

runtime: Option<String>

The runtime in which to run the function. Required when deploying a new function, optional when updating an existing function. For a complete list of possible choices, see the gcloud command reference.

service_account_email: Option<String>

The email of the function's service account. If empty, defaults to {project_id}@appspot.gserviceaccount.com.

source_archive_url: Option<String>

The Google Cloud Storage URL, starting with gs://, pointing to the zip archive which contains the function.

source_repository: Option<SourceRepository>

Beta Feature The source repository where a function is hosted.

source_token: Option<String>

Input only. An identifier for Firebase function sources. Disclaimer: This field is only supported for Firebase function deployments.

source_upload_url: Option<String>

The Google Cloud Storage signed URL used for source uploading, generated by google.cloud.functions.v1.GenerateUploadUrl

status: Option<String>

Output only. Status of the function deployment.

timeout: Option<String>

The function execution timeout. Execution is considered failed and can be terminated if the function is not completed at the end of the timeout period. Defaults to 60 seconds.

update_time: Option<String>

Output only. The last update timestamp of a Cloud Function.

version_id: Option<String>

Output only. The version identifier of the Cloud Function. Each deployment attempt results in a new version of a function being created.

vpc_connector: Option<String>

The VPC Network Connector that this cloud function can connect to. It can be either the fully-qualified URI, or the short name of the network connector resource. The format of this field is projects/*/locations/*/connectors/* This field is mutually exclusive with network field and will eventually replace it. See the VPC documentation for more information on connecting Cloud projects.

vpc_connector_egress_settings: Option<String>

The egress settings for the connector, controlling what traffic is diverted through it.