Struct google_cloudfunctions1::CloudFunction

pub struct CloudFunction {
    pub event_trigger: Option<EventTrigger>,
    pub status: Option<String>,
    pub update_time: Option<String>,
    pub description: Option<String>,
    pub max_instances: Option<i32>,
    pub source_repository: Option<SourceRepository>,
    pub https_trigger: Option<HttpsTrigger>,
    pub source_archive_url: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub version_id: Option<String>,
    pub entry_point: Option<String>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub available_memory_mb: Option<i32>,
    pub vpc_connector: Option<String>,
    pub environment_variables: Option<HashMap<String, String>>,
    pub source_upload_url: Option<String>,
    pub service_account_email: Option<String>,
    pub timeout: Option<String>,
    pub runtime: Option<String>,
}

Describes a Cloud Function that contains user computation executed in response to an event. It encapsulate function and triggers configurations. LINT.IfChange

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations functions patch projects (request)

locations functions get projects (response)

locations functions create projects (request)

Fields
event_trigger: Option<EventTrigger>

A source that fires events in response to a condition in another service.

status: Option<String>

Output only. Status of the function deployment.

update_time: Option<String>

Output only. The last update timestamp of a Cloud Function.

description: Option<String>

User-provided description of a function.

max_instances: Option<i32>

The limit on the maximum number of function instances that may coexist at a given time.

source_repository: Option<SourceRepository>

Beta Feature

The source repository where a function is hosted.

https_trigger: Option<HttpsTrigger>

An HTTPS endpoint type of source that can be triggered via URL.

source_archive_url: Option<String>

The Google Cloud Storage URL, starting with gs://, pointing to the zip archive which contains the function.

labels: Option<HashMap<String, String>>

Labels associated with this Cloud Function.

version_id: Option<String>

Output only. The version identifier of the Cloud Function. Each deployment attempt results in a new version of a function being created.

entry_point: Option<String>

The name of the function (as defined in source code) that will be executed. Defaults to the resource name suffix, if not specified. For backward compatibility, if function with given name is not found, then the system will try to use function named "function". For Node.js this is name of a function exported by the module specified in source_location.

name: Option<String>

A user-defined name of the function. Function names must be unique globally and match pattern projects/*/locations/*/functions/*

network: Option<String>

The VPC Network that this cloud function can connect to. It can be either the fully-qualified URI, or the short name of the network resource. If the short network name is used, the network must belong to the same project. Otherwise, it must belong to a project within the same organization. The format of this field is either projects/{project}/global/networks/{network} or {network}, where {project} is a project id where the network is defined, and {network} is the short name of the network.

This field is mutually exclusive with vpc_connector and will be replaced by it.

See the VPC documentation for more information on connecting Cloud projects.

This feature is currently in alpha, available only for whitelisted users.

available_memory_mb: Option<i32>

The amount of memory in MB available for a function. Defaults to 256MB.

vpc_connector: Option<String>

The VPC Network Connector that this cloud function can connect to. It can be either the fully-qualified URI, or the short name of the network connector resource. The format of this field is projects/*/locations/*/connectors/*

This field is mutually exclusive with network field and will eventually replace it.

See the VPC documentation for more information on connecting Cloud projects.

This feature is currently in alpha, available only for whitelisted users.

environment_variables: Option<HashMap<String, String>>

Environment variables that shall be available during function execution.

source_upload_url: Option<String>

The Google Cloud Storage signed URL used for source uploading, generated by google.cloud.functions.v1.GenerateUploadUrl

service_account_email: Option<String>

The email of the function's service account. If empty, defaults to {project_id}@appspot.gserviceaccount.com.

timeout: Option<String>

The function execution timeout. Execution is considered failed and can be terminated if the function is not completed at the end of the timeout period. Defaults to 60 seconds.

runtime: Option<String>

Required. The runtime in which the function is going to run. Choices:

nodejs6: Node.js 6

nodejs8: Node.js 8

nodejs10: Node.js 10

python37: Python 3.7

go111: Go 1.11